refactor: raceService 리펙터링
